# Improve performance of `is_energy_preserving` via sparse matrices