--- a/.github/workflows/FormatCheck.yml
+++ b/.github/workflows/FormatCheck.yml
@@ -36,6 +36,6 @@ jobs:
               exit(0)
           else
               @error "Some files have not been formatted !!!"
-              write(stdout, out)
+              println(out)
               exit(1)
           end'