--- a/Project.toml
+++ b/Project.toml
@@ -12,6 +12,7 @@ Polynomials = "f27b6e38-b328-58d1-80ce-0feddd5e7a45"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 RootedTrees = "47965b36-3f3e-11e9-0dcf-4570dfd42a8c"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [weakdeps]
 SymEngine = "123dc426-2d89-5057-bbad-38513e3affd8"