--- a/src/BSeries.jl
+++ b/src/BSeries.jl
@@ -1689,7 +1689,7 @@ function _is_energy_preserving(trees, coefficients)
     #       of the rank computation is based on `Float64`. Thus, it will usually
     #       not work with coefficients given only in 32 bit precision.
     if eltype(coefficients) <: Union{Float64, Rational{Int8}, Rational{Int16},
-        Rational{Int32}, Rational{Int64}, Rational{Int128}}
+             Rational{Int32}, Rational{Int64}, Rational{Int128}}
         # These types support efficient computations in sparse matrices
         _is_energy_preserving_sparse(trees, coefficients)
     else