--- a/src/BSeries.jl
+++ b/src/BSeries.jl
@@ -1688,8 +1688,12 @@ function _is_energy_preserving(trees, coefficients)
     # TODO: `Float32` would also be nice to have. However, the default tolerance
     #       of the rank computation is based on `Float64`. Thus, it will usually
     #       not work with coefficients given only in 32 bit precision.
-    if eltype(coefficients) <: Union{Float64, Rational{Int8}, Rational{Int16},
-             Rational{Int32}, Rational{Int64}, Rational{Int128}}
+    # For very few coefficients, dense operations are more efficient than
+    # sparse operations.
+    if (length(coefficients) > 20 &&
+        eltype(coefficients) <: Union{Float64,
+              Rational{Int8}, Rational{Int16}, Rational{Int32}, Rational{Int64},
+              Rational{Int128}})
         # These types support efficient computations in sparse matrices
         _is_energy_preserving_sparse(trees, coefficients)
     else